pyproject.toml
[project]
name = "vaultmind-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
    "sentence-transformers>=5.3.0",
    "watchdog>=6.0.0",
]

[project.scripts]
vaultmind = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
include = ["main.py", "vaultmind/"]
